Added AbstractFilesystemCheckService
